+++ VolquetesFront/src/app/app-layout/header/header.component.css
@@ -23,5 +23,5 @@ header {
 }
 
 .burger-button {
-  margin-left: auto;
+  margin-left: 2%;
 }
