+++ VolquetesFront/src/app/app-layout/header/header.component.html
@@ -1,6 +1,7 @@
 <main>
   <header>
-    <img src="../assets/logo.svg" alt="logo" class="logo">Volquetes
     <app-burger-button class="burger-button"></app-burger-button>
+    <img src="../assets/logo.svg" alt="logo" class="logo">Volquetes
+    
   </header>
 </main>
